move all status logging from PythonCalls to BiosimulationLog
