+++ vcell-cli/src/main/java/org/vcell/cli/run/SedmlJob.java
@@ -124,7 +124,7 @@ public class SedmlJob {
             this.sedmlName = sedmlNameSplit[sedmlNameSplit.length - 1];
             this.LOG_OMEX_MESSAGE.append("Processing ").append(this.sedmlName).append(". ");
             logger.info("Processing SED-ML: " + this.sedmlName);
-            PythonCalls.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.RUNNING, this.RESULTS_DIRECTORY_PATH);
+            BiosimulationLog.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.RUNNING, this.RESULTS_DIRECTORY_PATH);
 
             this.DOC_STATISTICS.setNumModels(sedmlFromOmex.getModels().size());
             for(Model m : sedmlFromOmex.getModels()) {
@@ -200,7 +200,7 @@ public class SedmlJob {
             Tracer.failure(e, prefix);
             this.reportProblem(e);
             this.somethingFailed = somethingDidFail();
-            PythonCalls.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.FAILED, this.RESULTS_DIRECTORY_PATH);
+            BiosimulationLog.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.FAILED, this.RESULTS_DIRECTORY_PATH);
             return false;
         } finally {
             if (span != null) {
@@ -346,8 +346,8 @@ public class SedmlJob {
         org.apache.commons.io.FileUtils.deleteDirectory(this.PLOTS_DIRECTORY);    // removing sedml dir which stages results.
 
         // Declare success!
-        PythonCalls.setOutputMessage(this.SEDML_LOCATION, this.sedmlName, this.RESULTS_DIRECTORY_PATH, "sedml", this.logDocumentMessage);
-        PythonCalls.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.SUCCEEDED, this.RESULTS_DIRECTORY_PATH);
+        BiosimulationLog.setOutputMessage(this.SEDML_LOCATION, this.sedmlName, this.RESULTS_DIRECTORY_PATH, "sedml", this.logDocumentMessage);
+        BiosimulationLog.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.SUCCEEDED, this.RESULTS_DIRECTORY_PATH);
         logger.info("SED-ML : " + this.sedmlName + " successfully completed");
         return true;
     }
@@ -383,7 +383,7 @@ public class SedmlJob {
         // We assume if no exception is returned that the plots pass
         for (Output output : this.sedml.getOutputs()){
             if (!(output instanceof Plot2D plot)) continue;
-            PythonCalls.updatePlotStatusYml(this.SEDML_LOCATION, plot.getId(), BiosimulationLog.Status.SUCCEEDED, this.RESULTS_DIRECTORY_PATH);
+            BiosimulationLog.updatePlotStatusYml(this.SEDML_LOCATION, plot.getId(), BiosimulationLog.Status.SUCCEEDED, this.RESULTS_DIRECTORY_PATH);
         }
     }
 
@@ -465,10 +465,10 @@ public class SedmlJob {
     private void reportProblem(Exception e) throws PythonStreamException, InterruptedException, IOException{
         logger.error(e.getMessage(), e);
         String type = e.getClass().getSimpleName();
-        PythonCalls.setOutputMessage(this.SEDML_LOCATION, this.sedmlName, this.RESULTS_DIRECTORY_PATH, "sedml", this.logDocumentMessage);
-        PythonCalls.setExceptionMessage(this.SEDML_LOCATION, this.sedmlName, this.RESULTS_DIRECTORY_PATH, "sedml", type, this.logDocumentError);
+        BiosimulationLog.setOutputMessage(this.SEDML_LOCATION, this.sedmlName, this.RESULTS_DIRECTORY_PATH, "sedml", this.logDocumentMessage);
+        BiosimulationLog.setExceptionMessage(this.SEDML_LOCATION, this.sedmlName, this.RESULTS_DIRECTORY_PATH, "sedml", type, this.logDocumentError);
         this.CLI_RECORDER.writeDetailedErrorList(e, this.BIOMODEL_BASE_NAME + ",  doc:    " + type + ": " + this.logDocumentError);
-        PythonCalls.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.FAILED, this.RESULTS_DIRECTORY_PATH);
+        BiosimulationLog.updateSedmlDocStatusYml(this.SEDML_LOCATION, BiosimulationLog.Status.FAILED, this.RESULTS_DIRECTORY_PATH);
     }
 
     private void recordRunDetails(SolverHandler solverHandler) throws IOException {
